MasterCaptainFile/requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.9.11
cffi==1.13.2
Click==7.0
colorama==0.4.1
cryptography==2.8
decouple==0.0.7
dj-database-url==0.5.0
Django==3.0.2
django-core==1.4.1
django-cors-headers==3.2.1
django-environ==0.4.5
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
Flask==1.1.1
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.0.0
pipenv==2018.11.26
psycopg2==2.8.4
pycparser==2.19
pylint==2.4.3
python-decouple==3.3
python-dotenv==0.10.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
Werkzeug==0.16.0
whitenoise==5.0.1
wrapt==1.11.2